requirements.txt
awscli
datasets
keras
keras_hub
loguru
matplotlib
pip
python-dotenv
rich
ruff
sacrebleu
sentencepiece
textual
tqdm
typer
-e .
